one more
I fixed this before but then git reverted it when I was renaming the file

+++ huggle/pythonengine.cpp
@@ -368,7 +368,7 @@ PythonScript *PythonEngine::PythonScriptObjFromPyObj(PyObject *object)
             return script;
         }
     }
-    HUGGLE_DEBUG("Unable to resolve script from script table, id: 0x" + QString::number(object), 4);
+    HUGGLE_DEBUG("Unable to resolve script from script table, id: 0x" + QString::number((int)object), 4);
     return nullptr;
 }
